#9 Add utils submodule for common constants, functions, and type hints
