+++ drs4/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ["ctrl", "daq", "specs"]
+__all__ = ["ctrl", "daq", "specs", "utils"]
 __version__ = "0.1.0"
 
 
@@ -6,3 +6,4 @@ __version__ = "0.1.0"
 from . import ctrl
 from . import daq
 from . import specs
+from . import utils
