+++ drs4/daq/tcp.py
@@ -2,20 +2,15 @@ __all__ = ["cross"]
 
 
 # standard library
-from collections.abc import Iterator, Sequence
-from contextlib import contextmanager
 from datetime import datetime, timezone
 from logging import getLogger
-from os import PathLike, getenv
+from os import getenv
 from pathlib import Path
-from tempfile import TemporaryDirectory
-from typing import Any, Literal as L, Optional, Union, get_args
+from typing import Optional, get_args
 
 
 # dependencies
-import numpy as np
 import xarray as xr
-from numpy.typing import NDArray
 from tqdm import tqdm
 from ..ctrl.self import ENV_CTRL_ADDR, ENV_CTRL_USER, run
 from ..specs.csv import TIME_FORMAT
@@ -29,12 +24,7 @@ from ..specs.zarr import (
     SideBand,
     open_csvs,
 )
-
-
-# type hints
-Axis = Optional[Union[Sequence[int], int]]
-Join = L["outer", "inner", "left", "right", "exact", "override"]
-StrPath = Union[PathLike[str], str]
+from ..utils import StrPath, XarrayJoin, set_workdir, unique
 
 
 # constants
@@ -63,7 +53,7 @@ def cross(
     # for file saving (optional)
     append: bool = False,
     integrate: bool = False,
-    join: Join = "inner",
+    join: XarrayJoin = "inner",
     overwrite: bool = False,
     progress: bool = False,
     workdir: Optional[StrPath] = None,
@@ -220,33 +210,3 @@ def cross(
             ds_if2.to_zarr(zarr_if2, mode="w")
 
         return zarr_if1.resolve(), zarr_if2.resolve()
-
-
-@contextmanager
-def set_workdir(workdir: Optional[StrPath] = None, /) -> Iterator[Path]:
-    """Set the working directory for output VDIF files."""
-    if workdir is not None:
-        yield Path(workdir).expanduser()
-    else:
-        with TemporaryDirectory() as workdir:
-            yield Path(workdir)
-
-
-def unique(array: NDArray[Any], /, axis: Axis = None) -> NDArray[Any]:
-    """Return unique values along given axis (axes)."""
-    if axis is None:
-        axis = list(range(array.ndim))
-
-    axes = np.atleast_1d(axis)
-    shape = np.array(array.shape)
-    newshape = np.prod(shape[axes]), *np.delete(shape, axes)
-
-    for ax in axes:
-        array = np.moveaxis(array, ax, 0)
-
-    result = np.unique(array.reshape(newshape), axis=0)
-
-    if result.shape[0] != 1:
-        raise ValueError("Array values are not unique.")
-
-    return result[0]
